[fix] fix tool result banner logics
