+++ client/src/components/relation/Visualization.js
@@ -5,7 +5,8 @@ import axios from 'axios';
 import { v4 as uuidv4 } from 'uuid';
 import options from './options'; 
 
-function Visualization() {
+function Visualization(props) {
+  const isDone = props.isDone
   useEffect(() => {
     const container = document.getElementById('graph-container');
 
@@ -43,7 +44,7 @@ function Visualization() {
     return () => {
       network.destroy();
     };
-  }, []);
+  }, [isDone]);
 
   return (
     <div>
